refactor: move main styles into frontend and adjust imports
