+++ battle-mode.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; style-src 'self'; script-src 'self'; media-src 'self'">
     <title>Kadir11 - Modo de Batalha</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="frontend/src/styles/main.css">
     <style>
         html {
             width: 100%;
@@ -28,7 +28,7 @@
             padding: 0;
         }
 
-        /* Título configurado em styles/main.css */
+        /* Título configurado em frontend/src/styles/main.css */
 
         #close-battle-mode {
             width: 20px;
@@ -46,7 +46,7 @@
             font-weight: bold;
             text-shadow: none;
         }
-        /* title-bar-buttons configurado em styles/main.css */
+        /* title-bar-buttons configurado em frontend/src/styles/main.css */
         #back-battle-mode {
             width: 20px;
             height: 20px;
