+++ create-pet.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; style-src 'self'; script-src 'self'; media-src 'self'">
     <title>Kadir11 - Criar Pet</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="frontend/src/styles/main.css">
     <style>
         html {
             width: 100%;
@@ -143,23 +143,23 @@
                 <h2>Escolha um Elemento</h2>
                 <div class="element-selection">
                     <button class="element-button" data-element="fogo">
-                        <img src="Assets/Elements/fogo.png" alt="Fogo" style="image-rendering: pixelated;">
+                        <img src="Assets/Elements/fogo.png" alt="Fogo">
                         <span class="element-label"></span>
                     </button>
                     <button class="element-button" data-element="ar">
-                        <img src="Assets/Elements/ar.png" alt="Ar" style="image-rendering: pixelated;">
+                        <img src="Assets/Elements/ar.png" alt="Ar">
                         <span class="element-label"></span>
                     </button>
                     <button class="element-button" data-element="terra">
-                        <img src="Assets/Elements/terra.png" alt="Terra" style="image-rendering: pixelated;">
+                        <img src="Assets/Elements/terra.png" alt="Terra">
                         <span class="element-label"></span>
                     </button>
                     <button class="element-button" data-element="agua">
-                        <img src="Assets/Elements/agua.png" alt="Água" style="image-rendering: pixelated;">
+                        <img src="Assets/Elements/agua.png" alt="Água">
                         <span class="element-label"></span>
                     </button>
                     <button class="element-button" data-element="puro">
-                        <img src="Assets/Elements/puro.png" alt="Puro" style="image-rendering: pixelated;">
+                        <img src="Assets/Elements/puro.png" alt="Puro">
                         <span class="element-label"></span>
                     </button>
                 </div>
