+++ frontend/src/components/BattleModeView.jsx
@@ -1,4 +1,4 @@
-import '../../styles/main.css'
+import '../styles/main.css'
 export default function BattleModeView() {
   return (
     <div className="window">
