+++ frontend/src/components/JourneySceneView.jsx
@@ -1,4 +1,4 @@
-import '../../styles/main.css'
+import '../styles/main.css'
 export default function JourneySceneView() {
   return (
     <div className="window">
