+++ frontend/src/components/LoadPetView.jsx
@@ -1,4 +1,4 @@
-import '../../styles/main.css'
+import '../styles/main.css'
 export default function LoadPetView() {
   return (
     <div className="window">
