+++ frontend/src/components/NestsView.jsx
@@ -1,4 +1,4 @@
-import '../../styles/main.css'
+import '../styles/main.css'
 export default function NestsView() {
   return (
     <div className="window">
