+++ frontend/src/components/StatusView.jsx
@@ -1,4 +1,4 @@
-import '../../styles/main.css'
+import '../styles/main.css'
 export default function StatusView() {
   return (
     <div className="window">
