+++ frontend/src/components/StoreView.jsx
@@ -1,4 +1,4 @@
-import '../../styles/main.css'
+import '../styles/main.css'
 export default function StoreView() {
   return (
     <div className="window">
