+++ frontend/src/components/TrainView.jsx
@@ -1,4 +1,4 @@
-import '../../styles/main.css'
+import '../styles/main.css'
 export default function TrainView() {
   return (
     <div className="window">
