+++ hatch.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; style-src 'self'; script-src 'self'; media-src 'self'">
     <title>Chocar Ovo</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="frontend/src/styles/main.css">
     <style>
         html, body { width:100%; height:100%; margin:0; padding:0; background:transparent; }
         #hatch-overlay {
