+++ journey-mode.html
@@ -5,14 +5,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; style-src 'self'; script-src 'self'; media-src 'self'">
     <title>Kadir11 - Modo Jornada</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="frontend/src/styles/main.css">
     <style>
         html { width: 100%; height: 100%; }
         body { background: transparent; margin: 0; padding: 0; }
         .window { position: relative; display: flex; flex-direction: column; margin: 0; padding: 0; width: fit-content; height: fit-content; }
-        /* Título configurado em styles/main.css */
+        /* Título configurado em frontend/src/styles/main.css */
         #close-journey-mode { width: 20px; height: 20px; background-color: #ff4444; border-radius: 50%; cursor: pointer; -webkit-app-region: no-drag; display: flex; align-items: center; justify-content: center; color: #813a3a; font-family: cursive; font-size: 12px; font-weight: bold; text-shadow: none; }
-        /* title-bar-buttons configurado em styles/main.css */
+        /* title-bar-buttons configurado em frontend/src/styles/main.css */
         #back-journey-mode { width: 20px; height: 20px; background-color: #44aaff; border-radius: 50%; cursor: pointer; display: flex; align-items: center; justify-content: center; color: #2a435b; font-family: cursive; font-size: 12px; font-weight: bold; text-shadow: none; }
         #journey-container { padding: 40px 10px 10px; margin: 5px; display: flex; flex-direction: column; width: fit-content; height: fit-content; }
