+++ journey-scene.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; style-src 'self'; script-src 'self'; media-src 'self'">
     <title>Cenário da Jornada</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="frontend/src/styles/main.css">
     <style>
         html, body {
             width: 100%;
@@ -25,7 +25,7 @@
             margin: 0;
             padding: 0;
         }
-        /* Título configurado em styles/main.css */
+        /* Título configurado em frontend/src/styles/main.css */
         #title-bar {
             position: relative;
             width: 100%;
@@ -58,7 +58,7 @@
             font-weight: bold;
             text-shadow: none;
         }
-        /* title-bar-buttons configurado em styles/main.css */
+        /* title-bar-buttons configurado em frontend/src/styles/main.css */
         #back-journey-scene {
             width: 20px;
             height: 20px;
