+++ start.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; style-src 'self' 'unsafe-inline'; script-src 'self' 'unsafe-eval' 'unsafe-inline'; media-src 'self'">
     <title>Kadir11 - Início</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="frontend/src/styles/main.css">
     <style>
         html {
             width: 100%;
@@ -148,10 +148,10 @@
 
 <body>
     <div class="window">
-        <img src="Assets/Logo/kadirnobg.png" alt="" style="image-rendering: pixelated;">
-        <img src="Assets/Logo/kadir11nme.png" alt="" class="animate-bg" style="image-rendering: pixelated;">
+        <img src="Assets/Logo/kadirnobg.png" alt="">
+        <img src="Assets/Logo/kadir11nme.png" alt="" class="animate-bg">
         <button class="button" id="start-button">Iniciar</button>
-        <button class="button" id="load-button" style="display: none;">Carregar</button>
+        <button class="button hidden" id="load-button">Carregar</button>
         <button class="button" id="exit-button">Sair</button>
         <button id="mute-button">🔊</button>
     </div>
