+++ store.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; style-src 'self'; script-src 'self'; media-src 'self'">
     <title>Kadir11 - Loja</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="frontend/src/styles/main.css">
     <style>
         html, body { width:100%; height:100%; margin:0; padding:0; background:transparent; }
         .window {
@@ -19,7 +19,7 @@
             padding: 0;
             margin: 0;
         }
-        /* Título configurado em styles/main.css */
+        /* Título configurado em frontend/src/styles/main.css */
         #close-store-window {
             width: 20px;
             height: 20px;
@@ -36,7 +36,7 @@
             font-weight: bold;
             text-shadow: none;
         }
-        /* title-bar-buttons configurado em styles/main.css */
+        /* title-bar-buttons configurado em frontend/src/styles/main.css */
         #back-store-window {
             width: 20px;
             height: 20px;
